+++ lib/c-api/src/wasm_c_api/unstable/mod.rs
@@ -1,1 +1,3 @@
+pub mod engine;
 pub mod module;
+pub mod target_lexicon;